move some files in preparation for test code change
