+++ cpp/include/cugraph/graph_generators.hpp
